test(reborn-itest): address multi-agent review (code-review + thermo-nuclear)
Structural / fidelity:
- AP1: delete the hand-mirrored `HarnessApprovalGateEvidence`; wire the REAL
  production `LocalDevApprovalGateEvidence` via a new test-support factory
  (`build_local_dev_approval_gate_evidence_for_test`) so the gate-evidence
  lookup can never drift from production.
- A1: move `assemble_thread_runtime` from builder.rs to group.rs (fields made
  pub(crate)); builder.rs drops 1092 -> 870 lines.
- A3: replace the `build_base()` ghost-tuple with a `GroupBaseData` struct.
- A4: `live_approvals` auto-approve disable is now fail-loud (expect/let-else
  unreachable) instead of an always-true `if let` that could vacuously pass.
- A5: remove the dead `GroupSharedStorage.storage` field.
- A2: extract the byte-identical `connect_google_account` OAuth helper into
  `tests/support/reborn/oauth_flow.rs` (shared via the existing #[path] tree);
  drop both local copies and the incorrect "can't be re-exported" comment.

Correctness / hygiene:
- SEC1: `assert_workspace_file_contains` reports only byte length on failure,
  never the full file contents (CI-log secret-leak surface).
- P1: poison-recovery on all `pending_approval_scopes` lock sites.
- LP3: move the `gate:approval-` prefix check into `submit_turn_until_blocked`;
  remove the duplicated inline checks from the approve/deny scenarios.
- Collapse thin auto-approve wrapper methods on HostRuntimeCapabilityHarness.

Coverage:
- T1: LibSql-backed approvals group test (gate/approve/deny/auto-approve-persist
  over a real on-disk backend, not just InMemory).
- T2: extension remove -> cross-thread search-absent scenario (uses "notion" to
  stay independent of scenario 1's "github"), with a non-vacuity guard.

Docs: probe-uniqueness comment, test_support.rs preamble, FixedCandidateSource
follow-up TODO, secrets libsql-gate rationale, and name the production call site
(build_local_runtime) on the extension-store test-support accessors.

Deferred (with rationale): unify the two harness.rs host-runtime constructors
(the tracked harness_mcp.rs extraction is the real fix), split test_support.rs
(806 < 1500 threshold), FixedCandidateSource real-enumeration test (refresh path
already covered at full fidelity; tracked by TODO).
