+++ crates/ironclaw_reborn_composition/src/extension_lifecycle.rs
@@ -177,9 +177,9 @@ impl RebornLocalExtensionManagementPort {
 
     /// Test-support access to the extension installation store.
     ///
-    /// Mirrors the `installation_store` field the production factory wires in
-    /// when composing `RebornLocalExtensionManagementPort`. For tests only —
-    /// zero bytes shipped in production builds.
+    /// Mirrors the `installation_store` field that `build_local_runtime` wires
+    /// in when constructing `RebornLocalExtensionManagementPort`. For tests
+    /// only — zero bytes shipped in production builds.
     #[cfg(feature = "test-support")]
     pub(crate) fn installation_store_for_test(&self) -> Arc<dyn ExtensionInstallationStore> {
         Arc::clone(&self.installation_store)
