+++ tests/reborn_group_approvals/scenario_gate_then_approve.rs
@@ -29,9 +29,6 @@ pub async fn run(g: &RebornIntegrationGroup) -> HarnessResult<()> {
     let (run_id, gate_ref) = h
         .submit_turn_until_blocked("write the approval file")
         .await?;
-    if !gate_ref.as_str().starts_with("gate:approval-") {
-        return Err(format!("expected a local-dev approval gate, got {gate_ref:?}").into());
-    }
 
     // Approve through the real resolver + resume; the gated capability re-runs.
     h.approve_gate(run_id, &gate_ref).await?;
